Fully realized DAO for employee and department, works in two sides, tested.

+++ src/main/java/dao/EmployeeDAO.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface EmployeeDAO {
     public boolean insertEmployee(Employee employee);
-    public Employee findEmployee(int key);
+    public Employee findEmployee(long key);
     public boolean updateEmployee(Employee employee);
-    public boolean deleteEmployee(int key);
+    public boolean deleteEmployee(long key);
     public List<Employee> getAllEmployees();
 }
